requirements.txt
aspose-words==23.7.0
Pillow==10.0.0